Migrate PendingActions to useNotificationData
